refactor : long->>decimal(deposit, withdrwal, amount), 빠진 어노테이션 추가
- SQS 메시지 수신 시 handleMessage()가 정상적으로 AccountService.createAccount()를 호출하는지 검증
- ObjectMapper를 이용한 JSON 파싱 및 예외 상황 처리 확인

+++ src/main/java/com/example/mockbank/application/dto/WithdrawRequest.java
@@ -5,11 +5,13 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class WithdrawRequest {
     @NotNull(message = "입금 금액은 필수입니다.")
     @Min(value = 1, message = "입금 금액은 1원 이상이어야 합니다.")
-    private Long amount;
+    private BigDecimal amount;
     private String memo; // 선택: 출금 메모
 }
